% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chile.R
\docType{data}
\name{chile}
\alias{chile}
\title{Chile Subduction Zone Data}
\format{An object of class \code{"data.frame"}}
\source{
\href https://earthquake.usgs.gov/earthquakes/search/
}
\usage{
data(chile)
}
\description{
Data from USGS https://earthquake.usgs.gov/earthquakes/search/ for a 
region along the Chilean coastline from a date range of 1900 - 2017.
}
\examples{
data(chile)
adj_beniplot(data = chile, strike = 13)


}
\references{
https://earthquake.usgs.gov/earthquakes/search/
}
\keyword{datasets}
