% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create Event Labels}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A character vector containing the columns "location', 'mag' and
'total_deaths'.}
}
\value{
A data frame with HTML-formatted label information.
}
\description{
Helper function that creates a new information column composed of:
date (in %Y-%m-%d format), magnitude and total number of casualties.
The resulting information are newline separated to keep the labels
adequately small. The labels do not show missing values. Instead, the respective
information is omitted.
}
\examples{
#create label column
southamerica \%>\%
filter(country == "CHILE" & year(date) >= 1970) \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  head()

# use labels in a plot
southamerica \%>\%
filter(country == "CHILE" & year(date) >= 1970) \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")

}
