% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Draws a Map with Earthquake Locations}
\usage{
eq_map(
  data,
  annot_col,
  scaling_factor = 1,
  color = "darkred",
  provider = "Esri.WorldTopoMap"
)
}
\arguments{
\item{data}{A data frame containing columns for "country",
"magnitude", "longitude" and "latitude". The latter two must be values in [deg].}

\item{annot_col}{A vector of values which will be used for labels.}

\item{scaling_factor}{A numeric value which scales the point size.
Defaults to 1.}

\item{color}{Character value specifying the color of the Earthquake locations.
Defaults to "darkred".}

\item{provider}{Name of the map provider. See: https://leaflet-extras.github.io/leaflet-providers/preview/) for available options.
Defaults to "Esri.WorldTopoMap".}
}
\value{
An interactive leaflet map.
}
\description{
\code{eq_map} plots the locations of earthquake events on a geographical map
    of one or multiple countries. Earthquake locations are plotted as
    red circles with labels and size scaled based on magnitude.
}
\examples{
library(dplyr)
# Plot earthquake events for a single country
southamerica \%>\%
  dplyr::filter(country == "CHILE" & lubridate::year(date) >= 1970) \%>\%
  eq_map(annot_col = "date", scaling_factor = 5)

# Plot earthquake events for multiple countries
southamerica \%>\%
  dplyr::filter(lubridate::year(date) > 2000) \%>\%
  eq_map(annot_col = "mag", scaling_factor = 5)

}
