% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoord_km.R
\name{geocoord_km}
\alias{geocoord_km}
\title{Convert Lat/Lon to Kilometers}
\usage{
geocoord_km(data)
}
\arguments{
\item{data}{a dataframe containing at least two vectors with geocoordinates labeled longitute and latitude}
}
\value{
A dataframe is printed to the console containing new cartesian (x, y) values in kilometers along with the original latitude and longitude values.
}
\description{
\code{geocoord_km} converts geocoordinates of latitude and longitude to kilometers. Projections of lat and lon are turned into (x,y) points centered around their mean in kilometers. Code from http://wiki.cbr.washington.edu/qerm/index.php/R/Converting_Geographic_Coordinates_to_Kilometers was used in making the function. See link for more information on calculation.
}
\examples{
geocoord_km(data = dataframe)
}
