% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{geom_timeline}
\alias{geom_timeline}
\alias{GeomTimeline}
\title{Plots a timeline of earthquake events}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{x}{A date vector specifying the occurrences of earthquakes.}

\item{xmin}{A date value specifying the left boundary of the timeline.}

\item{xmax}{A date value specifying the right boundary of the timeline.}

\item{y}{A factor specifying any stratification for multiple timelines.
By default one timeline is plotted in the middle of the viewport.}
}
\value{
A layer \code{ggproto} object.
}
\description{
Creates a timeline geom for ggplot2, which can be used to visualize
events or intervals over a continuous time axis.
}
\section{Aesthetics}{
 `geom_timeline()` understands the following
  aesthetics (required aesthetics are in bold):
  \itemize{\item{**`x`**} \item{**`xmin`**} \item{**`xmax`**}
  \item{`y`} \item{`linewidth`} \item{`linetype`} \item{`linecolour`}}
  \item{`alpha`} \item{`colour`} \item{`shape`} \item{`size`}
  \item{`stroke`}
}

\examples{
#single timeline without `y` aesthetic
p <- southamerica \%>\%
filter(country == "CHILE" & year(date) >= 1970) \%>\%
 ggplot(aes(x = date)))

p + geom_timeline(aes(xmin = min(date), xmax = max(date))

#multiple timelines
s <- southamerica \%>\%
 filter(lubridate::year(date) >= 1990) \%>\%
 ggplot(aes(x = date, y = country))

s + geom_timeline(aes(xmin = min(date), xmax = max(date))

}
\keyword{datasets}
