% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{geom_timeline_label}
\alias{geom_timeline_label}
\alias{GeomTimelineLabel}
\title{Add labels to geom_timeline}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  n_max = NULL,
  limit_var = NULL,
  label_dodge = FALSE,
  check_overlap = FALSE,
  ...,
  inherit.aes = TRUE
)
}
\arguments{
\item{n_max}{An integer value that limits the number of
labels based on the values mapped to `limit_var`. If not defined all labels
are displayed. Defaults to `NULL`.}

\item{limit_var}{A variable used in conjunction with `n_max` to limit
the number of displayed labels to the equal `n_max`, based on the highest
values of the `limit_var`. Defaults to `NULL`.}

\item{label_dodge}{Logical toggles vertically alternating orientation of
labels to reduce overlapping. Defaults to `FALSE`.}

\item{x}{A date vector specifying x values.}

\item{label}{A vector specifying the information to be displayed.}
}
\value{
A layer \code{ggproto} object.
}
\description{
This function adds label to the timeline geom. The distance between the timeline
and the labels is static when no `size` aesthetic is specified and calculated
based on max point size otherwise.
}
\section{Aesthetics}{
 `geom_timeline_label()` understands the following
  aesthetics (required aesthetics are in bold):
  \itemize{\item{**`x`**} \item{**`label`**}
  \item{`linewidth`} \item{`linetype`} \item{`linecolour`}}
  \item{`alpha`} \item{`colour`} \item{`shape`} \item{`size`}
  \item{`stroke`}
}

\examples{
p <- southamerica \%>\%
filter(country == "CHILE" & year(date) >= 1970) \%>\%
 ggplot(aes(x = date)) +
 geom_timeline(aes(
  xmin = min(date),
  xmax = max(date)
  )
 )

p + geom_timeline_label(aes(label = location))

#use `label_dodge` to reduce overlapping
p + geom_timeline_label(aes(label = location), label_dodge = TRUE)

#limit number of displayed labels based on earthquake magnitude
p + geom_timeline_label(aes(label = mag))
p + geom_timeline_label(aes(label = mag, limit_var = mag), n_max = 3)

}
\keyword{datasets}
