% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_transform.R
\name{polar_transform}
\alias{polar_transform}
\title{Rotate cartesian coordinates}
\usage{
polar_transform(data, rotation = 0, plots = TRUE)
}
\arguments{
\item{data}{a dataframe containing cartesian coordinates labeled x and y}

\item{rotation}{The desired value of rotation in degrees}

\item{plots}{if \code{plots = TRUE} a plot showing both the original x, y data and the rotated data (in green) is printed. If \code{plots = FALSE}, a plot is not printed.}
}
\value{
A dataframe is printed to the console containing original cartesian (x, y) values and the new rotated (new.x, new.y) values. If \code{plots = TRUE} a plot showing both the original and new coordinates (colored green) is plotted
}
\description{
\code{polar_transform} rotates cartesian coordinates about the origin at a given degree
}
\examples{
polar_transform(data = dataframe, rotation = 45, plots = TRUE)
}
