% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southamerica.R
\docType{data}
\name{southamerica}
\alias{southamerica}
\title{NCEI earthquake data}
\format{

}
\source{
<https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search>
}
\usage{
southamerica
}
\description{
A subset of data from the National Centers for Environmental Information
earthquake database from 05/2024. The dataset contains earthquake data
for the south american continent starting from 1900 and has been filtered
for complete observations.
}
\keyword{datasets}
