% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create Event Labels}
\usage{
eq_create_label(data, defaults = TRUE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{defaults}{Specifies whether the default variables 'location', 'mag'
and 'total_deaths' are to be used. Defaults to `TRUE`. If `FALSE` and
no additional variables are added via `...` the labell will be empty.}

\item{...}{Additional variables to add to the label or instead of defaults.}
}
\value{
A data frame with HTML-formatted label information.
}
\description{
Helper function that creates an HMTL-formatted information
    based on user-specified variables. By default uses 'location','mag' and '
    total_deaths' variables.
}
\examples{
library(dplyr)

#create label column
southamerica \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  head()

# use labels in a plot
southamerica \%>\%
  filter(country == "CHILE" & lubridate::year(date) >= 1970) \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")

}
