% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{geom_timeline_label}
\alias{geom_timeline_label}
\alias{GeomTimelineLabel}
\title{Add labels to geom_timeline}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  n_max = NULL,
  limit_var = NULL,
  label_dodge = FALSE,
  check_overlap = FALSE,
  ...,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{n_max}{An integer value that limits the number of
labels based on the values mapped to `limit_var`. If not defined all labels
are displayed. Defaults to `NULL`.}

\item{limit_var}{A variable used in conjunction with `n_max` to limit
the number of displayed labels to the equal `n_max`, based on the highest
values of the `limit_var`. Defaults to `NULL`.}

\item{label_dodge}{Logical toggles vertically alternating orientation of
labels to reduce overlapping. Defaults to `FALSE`.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{x}{A date vector specifying x values.}

\item{label}{A vector specifying the information to be displayed.}
}
\value{
A layer \code{ggproto} object.
}
\description{
This function adds label to the timeline geom. The distance between the timeline
and the labels is static when no `size` aesthetic is specified and calculated
based on max point size otherwise.
}
\section{Aesthetics}{
 `geom_timeline_label()` understands the following
  aesthetics (required aesthetics are in bold):
  \itemize{\item{**`x`**} \item{**`label`**}
  \item{`linewidth`} \item{`linetype`} \item{`linecolour`}
  \item{`alpha`} \item{`colour`} \item{`shape`} \item{`size`}
  \item{`stroke`}}
}

\examples{
p <- southamerica \%>\%
filter(country == "CHILE" & year(date) >= 1970) \%>\%
 ggplot(aes(x = date)) +
 geom_timeline(aes(
  xmin = min(date),
  xmax = max(date)
  )
 )

p + geom_timeline_label(aes(label = location))

#use `label_dodge` to reduce overlapping
p + geom_timeline_label(aes(label = location), label_dodge = TRUE)

#limit number of displayed labels based on earthquake magnitude
p + geom_timeline_label(aes(label = mag))
p + geom_timeline_label(aes(label = mag, limit_var = mag), n_max = 3)

}
\keyword{datasets}
